[fix] configuredAudioSession() 앱실행시 한 번만
