+++ alsongDalsong/ASAudioKit/ASAudioKit/ASAudioPlayer/ASAudioPlayer.swift
@@ -21,7 +21,6 @@ public actor ASAudioPlayer: NSObject {
     /// 녹음파일을 재생하고 옵션에 따라 재생시간을 설정합니다.
     public func startPlaying(data: Data, option: PlayType = .full, fade _: Bool = false, isLoop: Bool = false) throws {
         do {
-            try configureAudioSession()
             audioPlayer = try AVAudioPlayer(data: data)
             audioPlayer?.delegate = self
             audioPlayer?.prepareToPlay()
@@ -87,18 +86,6 @@ public actor ASAudioPlayer: NSObject {
 
     public func setOnPlaybackFinished(_ handler: @Sendable @escaping () async -> Void) {
         onPlaybackFinished = handler
-    }
-
-    private func configureAudioSession() throws {
-        do {
-            let session = AVAudioSession.sharedInstance()
-            try session.setCategory(.playAndRecord, mode: .default, options: [.defaultToSpeaker, .mixWithOthers])
-            try session.setActive(true, options: .notifyOthersOnDeactivation)
-        } catch {
-            // TODO: 세션 설정 실패에 따른 처리
-            ErrorHandler.handle(error)
-            throw ASAudioError.configureAudioSession
-        }
     }
 }
